Use DeepReactive in dialog example
