+++ README.md
@@ -8,16 +8,16 @@ This repository contains examples of how to build application functionalities in
 2. [Create a JSON editor for a Pydantic model](examples/pydantic_monaco)
 3. [Conditionally rendering elements](examples/conditional_rendering)
 4. [Conditionally disabling elements](examples/conditional_disabling)
-5. [Changing Pydantic rules based on user input](examples/dynamic_pydantic_rules)
-6. [Complex Pydantic rules](examples/complex_pydantic_rules)
-7. [Selecting datafiles from the server](examples/data_selector)
-8. [Running a Galaxy tool](examples/run_galaxy_tool)
-9. [Running a Galaxy workflow](examples/run_galaxy_workflow)
-10. [Working with Plotly](examples/plotly)
-11. [Working with Matplotlib](examples/matplotlib)
-12. [Working with VTK](examples/vtk)
-13. [Synchronizing changes between tabs](examples/multitab)
-14. [Manually syncing the view and model](examples/trame_state)
+5. [Creating a dialog](examples/dialog)
+6. [Changing Pydantic rules based on user input](examples/dynamic_pydantic_rules)
+7. [Complex Pydantic rules](examples/complex_pydantic_rules)
+8. [Selecting datafiles from the server](examples/data_selector)
+9. [Running a Galaxy tool](examples/run_galaxy_tool)
+10. [Running a Galaxy workflow](examples/run_galaxy_workflow)
+11. [Working with Plotly](examples/plotly)
+12. [Working with Matplotlib](examples/matplotlib)
+13. [Working with VTK](examples/vtk)
+14. [Synchronizing changes between tabs](examples/multitab)
 
 We also provide examples that take advantage of ORNL resources:
